+++ CHANGELOG.md
@@ -13,6 +13,7 @@ MolSimToolkit.jl Changelog
 Version 1.33.0-DEV
 --------------
 - ![FEATURE][badge-feature] Introduce the option to compute the RMSD of a structure, given the alignment of other structure (or parts of); i. e. the `rmsd_indices` option of `rmsd`. 
+- ![FEATURE][badge-feature] Support string selections in `rmsd` and `rmsd_matrix` functions.
 
 Version 1.32.3
 --------------